Fix most clang warnings

+++ include/texture.inl
@@ -5,8 +5,7 @@ std::vector<T> Texture::get_pixels(GLenum type, GLenum format, size_t size, GLin
     vec.reserve(size);
     bind(Tex2D);
     glGetTexImage(Tex2D, level, format, type, buf);
-    for(int i = 0 ; i != size ; ++i) {
-        
+    for(size_t i = 0 ; i != size ; ++i) {
     }
     return vec;
 }
